pyproject.toml
[tool.poetry]
name = "backendflask"
version = "0.1.0"
description = ""
authors = ["Aurelio Marques Vulcão <vulcao@vulcaotech.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask_pymongo = "^2.3.0"
flask-cors = "^3.0.8"
pymongo = "^3.10.1"
dnspython = "^1.16.0"
python-decouple = "^3.3"
pep8 = "^1.7.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
